% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chirstmas_Tree.R
\name{christ_tree}
\alias{christ_tree}
\title{Draw a Christmas Tree}
\usage{
christ_tree(bg = "#FDF5DC", mcol = "gold", tcol = "#256F00")
}
\arguments{
\item{bg}{color of canvas}

\item{mcol}{color of title "Merry Christmas"}

\item{tcol}{color of "for you"}
}
\value{
a christmas tree
}
\description{
Draw a Christmas Tree
}
\examples{
christ_tree(bg = 'grey', mcol = 'black', tcol = 'blue')

}
