% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/ff16.R
\name{FF16_Individual}
\alias{FF16_Individual}
\alias{FF16_Parameters}
\title{Create a FF16 Plant or Node}
\usage{
FF16_Individual(s = FF16_Strategy())

FF16_Parameters()
}
\arguments{
\item{s}{A \code{\link{FF16_Strategy}} object}
}
\description{
Create a FF16 Plant or Node
}
\examples{
pl <- FF16_Individual()
pl$height
}
