% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{FF16_Environment}
\alias{FF16_Environment}
\alias{K93_Environment}
\alias{TF24_Environment}
\title{Create an FF16_Environment object.}
\usage{
FF16_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths
)

K93_Environment()

TF24_Environment(
  light_availability_spline_rescale_usually,
  soil_number_of_depths
)
}
\arguments{
\item{light_availability_spline_rescale_usually}{whether to rescale intervals when estimating light environment}

\item{soil_number_of_depths}{Number of soil layers to include}
}
\description{
Create an FF16_Environment object.
}
