% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16r.R
\name{FF16r_Individual}
\alias{FF16r_Individual}
\title{Create a FF16r Plant or Node}
\usage{
FF16r_Individual(s = FF16r_Strategy())
}
\arguments{
\item{s}{A \code{\link{FF16r_Strategy}} object}
}
\description{
Create a FF16r Plant or Node
}
\examples{
pl <- FF16r_Individual()
pl$height
}
