% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff16w.R
\name{FF16w_Individual}
\alias{FF16w_Individual}
\alias{FF16w_Parameters}
\title{Create a FF16w Plant or Node}
\usage{
FF16w_Individual(s = FF16w_Strategy())

FF16w_Parameters()
}
\arguments{
\item{s}{A \code{\link{FF16w_Strategy}} object}
}
\description{
Create a FF16w Plant or Node
}
\examples{
pl <- FF16w_Individual()
pl$height
}
