% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{Leaf}
\alias{Leaf}
\title{Leaf model}
\usage{
Leaf(
  vcmax_25,
  c,
  b,
  psi_crit,
  beta2,
  jmax_25,
  hk_s,
  a,
  curv_fact_elec_trans,
  curv_fact_colim,
  GSS_tol_abs,
  vulnerability_curve_ncontrol,
  ci_abs_tol,
  ci_niter
)
}
\arguments{
\item{vcmax_25}{rubisco rate}
}
\description{
Leaf model
}
