\name{na.pass}
\alias{na.pass}
\title{Pass through missing values}
\description{
  A \code{na.action} methods that does nothing.
}
\usage{
na.pass(x)
}
\arguments{
 \item{x}{any \R object}
}
\value{
  \code{x}, unchanged.
}
\seealso{\code{\link{na.omit}},  \code{\link{na.fail}},
  \code{\link{na.tree.replace}}}
}
\keyword{tree}
