% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf24.R
\name{TF24_Individual}
\alias{TF24_Individual}
\title{Create a TF24 Plant or Node}
\usage{
TF24_Individual(s = TF24_Strategy())
}
\arguments{
\item{s}{A \code{\link{TF24_Strategy}} object}
}
\description{
Create a TF24 Plant or Node
}
\examples{
pl <- TF24_Individual()
pl$height
}
