\name{tree-internal}
\alias{labels.tree}
\alias{model.frame.tree}
\title{Internal Functions Used by the `tree' Package}
\description{
  Internal functions used in the \code{tree} package.
}
\usage{
labels.tree(object, pretty = TRUE, collapse = TRUE, ...)
model.frame.tree(formula, ...)

}
\note{These are not intended for use by end users.}
\keyword{internal}
