% This file is part of the 'tree' package for R
% Copyright (C) B. D. Ripley 1998-2007
% It is distributed under the GPL version 2 or 3

\name{tree-internal}
\alias{labels.tree}
\alias{model.frame.tree}
\title{Internal Functions Used by the 'tree' Package}
\description{
  Internal functions used in the \pkg{tree} package.
}
\usage{
\method{labels}{tree}(object, pretty = TRUE, collapse = TRUE, \dots)

\method{model}{frame.tree}(formula, \dots)
}
\note{These are not intended for use by end users.}
\keyword{internal}
